+++ src/search/refinement/SimpleRefinementOperator.java
@@ -6,12 +6,13 @@ import arff.attribute.Constraint;
 import group.Group;
 import search.quality.AbstractQualityMeasure;
 
+import java.math.BigInteger;
 import java.util.HashSet;
 import java.util.Set;
 
 public class SimpleRefinementOperator extends AbstractRefinementOperator {
     @Override
-    public Set<Group> generate(Group seed, Dataset dataset, AbstractQualityMeasure qualityMeasure, HashSet<Long> encounteredGroups, HashSet<String> blacklist) {
+    public Set<Group> generate(Group seed, Dataset dataset, AbstractQualityMeasure qualityMeasure, HashSet<BigInteger> encounteredGroups, HashSet<String> blacklist) {
         HashSet<Group> groups = new HashSet<>();
         //Extend the seed by attributes and constraints that are not similar.
